fixed gaspump collisions
gas pump's origin was upper left corner which caused the collision
normal to be flipped in certain circumstances (collisions near its
origin)
needed to move the origin to the center of the sprite to rectify.

would be nice if I had a cleaner way of doing that -- define the width
and height and the image width/height/offset and where the origin is.