fixed the blipping out of existance bug!
The collision normal gets scaled by the collision depth and occasionally
that depth is exactly zero giving us a [0,0] normal. When this happened
the resolveCollision() code does a division by zero making the object
velocities NaN. Position gets set by the velocity making it NaN and we
use that to try and find the current node which turns up empty. No
current node, we assume the sprite is off the screen, and we call it's
die() method.

Yuck. So just have to check for zero before scaling the collision
normal.