fix blorping bug
new Hondas were created with the same base Car object so when one was
loaded, all of a sudden the dude was driving two cars, one of which was
past the section load threshold cauisng another load. Then the first one
would be past the threshold causing another load, and on and on.